refactor: reorganize folders and move elements to modules
