+++ src/app.module.ts
@@ -1,10 +1,10 @@
 import { Module, MiddlewareConsumer } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { ImagesModule } from './images/images.module';
+import { ImagesModule } from './modules/images/images.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Image } from './images/entities/image.entity';
-import { FirebaseModule } from './firebase/firebase.module';
+import { Image } from './modules/images/entities/image.entity';
+import { FirebaseModule } from './modules/firebase/firebase.module';
 import logger from './logger';
 @Module({
   imports: [
@@ -24,7 +24,7 @@ import logger from './logger';
   controllers: [AppController],
   providers: [AppService],
 })
-export class AppModule { 
+export class AppModule {
   configure(consumer: MiddlewareConsumer) {
     consumer
       .apply((req, res, next) => {
