+++ src/modules/images/dto/upload-image.dto.ts
