+++ src/modules/images/entities/image.entity.ts
